+++ apps/react-vite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # react-vite
 
+## 0.0.1-beta.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [4841960]
+- Updated dependencies
+  - @basictech/react@0.2.0-beta.3
+
 ## 0.0.1-beta.2
 
 ### Patch Changes